Nearly finished implementation of the new messsage service (minus proper formatting of the code using the code standard):
This implementation for mysql contains a few optimizations w.r.t.
the previous implementation:

-use of buffers for message queue to prevent single row
inserts in large tables
-table that specifies if there is a message in the queue for a component
to prevent unecessary queries on a large message queue
-mutli queue option (can be turned of/on): each component creates his own tables,
based on the name it registers with and messages for this component are stored
in these tables, thereby keeping the queue length managable. This option
creates 6 tables per component (e.g. 10 components create 60 tables)
-priority messages: separate tables whos messages are examined before
the normal messages.
-periodic purging of history queue with option to keep always the last
<x> messages in the history queue.

git-svn-id: svn+ssh://svn.cern.ch/reps/CMSDMWM/WMCore/trunk@2898 4525493e-7705-40b1-a816-d608a930855b