Implementation of a persistent threadpool. Payloads (pickled objects) not handled by a thread are stored in a queue with import output buffers.
The Slave objects (that run in the thread) are created on demand to prevent
a large startup time upfront and resued when possible until the maximum
number of threads is reached.

Information on threasds can be stored in a single queue (all threadpools
share that), or in a multi queue (one queu per pool). The latter generates
more tables.

git-svn-id: svn+ssh://svn.cern.ch/reps/CMSDMWM/WMCore/trunk@2910 4525493e-7705-40b1-a816-d608a930855b