-Added some additonal information at WMFactory exception message when loading an object
-Added query to MsgService that counts how many messages for a component
are in the buffers and queues. If 0 then we know we do not have to query
these queues until a new one arrives.

-Moved part of the Harness constructor to an initInThread method.
This method is called during startup. It enables to startup the component
as a thread while still using the threadlocal attributes, which is
useful for testing.

-Updated the ErrorHandler and TestComponent so their initialization
takes place in the preInitialization method instead of the constructor.

-Propagated the procid of the message service object created in the mainthread,
to the thread slaves so when a message is published in a thread, its source
is the component in which this thread is running.

git-svn-id: svn+ssh://svn.cern.ch/reps/CMSDMWM/WMCore/trunk@2990 4525493e-7705-40b1-a816-d608a930855b