-Added __init__ files in several tests so they can be used in suite testing
-Move/renamed some files in the WMCORE/standards directory.

-Updated the WMCORE/standards/README file to include various steps
when developing libraries.

-Setup project wide tests (WMCORE/standards/test_code and
WMCORE/standards/test_style which test the overal style and tests.
After which a report is generated. The scripts contain a mapping from the
module or test to the developer who is responsible to he/she can be
contacted. Eventually you want to have some automated email being send.

-Added a module WMQuality that stores the generics for testing. This module
can than also be used in other projects for quality testing

git-svn-id: svn+ssh://svn.cern.ch/reps/CMSDMWM/WMCore/trunk@3053 4525493e-7705-40b1-a816-d608a930855b