I modified the interfaces of 2 methods in the WMInit class:
def setLogging(self,logFile = None, logName = None, logLevel = logging.INFO, logExists = True):
def settabaseConnection(self, dbConfig, dialect = None, socketLoc = None, flavor = 'ProdAgent'):

The first accomodates the case where an external project can just call this method without
any parameters, but still has the flexibility to be used in the TestInit class where we want
to associate tests with certain (named) log files

The second enables to pass on just a dictionary. Currently this is geared to ProdAgent
(which is the default flavor), but if there are multiple projects with different formats
(other than the database string), it can be accomodated too.

These and the other methods are offered as auxilary methods, and projects such as
proodagent need to choose what they want to use, and also decide within their project
if they want to bundle all methods into one (e.g. inherit from this class and create
a doItAll(parameters) method, and perhaps even a single method in a file that is called
with just a single import.

git-svn-id: svn+ssh://svn.cern.ch/reps/CMSDMWM/WMCore/trunk@3377 4525493e-7705-40b1-a816-d608a930855b