Added a feature to the harness to alternate between sending messages through bypassing the message service and using the message service.
E.g. handleMessage(type='', payload='') will try and get a message from the message queue
if type is an empty string. Otherwise it will try and handle the message passed to it.

This is comes in handy if the handler for one component generates a message for another compnent and you
want to test the handler without having to start the component.

git-svn-id: svn+ssh://svn.cern.ch/reps/CMSDMWM/WMCore/trunk@3820 4525493e-7705-40b1-a816-d608a930855b