Add the CouchDB API to CVS.
The API is a set of known REST interactions. This is generalised in Requests and made specific for CouchDB in CouchRequests. CouchRequests is subclassed by CouchServer (representing the server itself) and Database (representing a database in the CouchServer). Document is provided as a conveniance for deleting documents, and is adictionary.

git-svn-id: svn+ssh://svn.cern.ch/reps/CMSDMWM/WMCore/trunk@3906 4525493e-7705-40b1-a816-d608a930855b