Adding SandboxCreator
SandboxCreator has one method which produces a sandbox, tars it up and returns the filename (as well as sticking it in the WMWorkload object). The tests on the SandboxCreator will fail without the __eq__ function I added to ConfigurationSection, but I'm going to wait for someone else's approval before I play with that.

git-svn-id: svn+ssh://svn.cern.ch/reps/CMSDMWM/WMCore/trunk@4311 4525493e-7705-40b1-a816-d608a930855b