Changed semantics around committing to CouchDB
Previously, the Database.commit() function had two different behaviors depending
on if there were documents enqueued to be submitted. It used the Bulk Commit
api if there was > 1 document to be submitted and the regular POST api
otherwise. Unfortunately, the two apis have completely different return values
which makes error testing much more difficult than it needs to be (you don't
explicitly choose whether or not you're submitting in bulk or not, so what you
expect to happen might not if a previous function didn't commit() for whatever
reason).

So, I've broken up commit into two different functions. commit and commitQueued
commit has the same behavior as before, it returns a dict with the id, rev and ok
status for the (one) document you pass to it. commitQueued dumps out the current
_queue from the Database object and returns:
( [list of good commits' id, rev, ok], [list of errored commits])

The client is still responsible for resubmitting any errored commits (conflicting updates, etc)

git-svn-id: svn+ssh://svn.cern.ch/reps/CMSDMWM/WMCore/trunk@4427 4525493e-7705-40b1-a816-d608a930855b