Initial checkin of a JobSubmitter draft.  This should take a jobGroup and out of it submit all the jobs within it using BossLite.  It depends on the BossLite code in PRODCOMMON at the moment.  It also does not work yet with Oracle.
This is a VERY preliminary draft, and has only been tested submitting to dummy
sites.  It is being committed now so that A) I don't lose it, B) other people can
now easily look at it, and C) so that we have a global record of how to handle
submissions to BossLite.

TODO: Bulk submit, and testing with a full job setup.

git-svn-id: svn+ssh://svn.cern.ch/reps/CMSDMWM/WMCore/trunk@4462 4525493e-7705-40b1-a816-d608a930855b