Fixed ChangeState. Apparently the 'none' state manages to break something between python and the database. I should change that tomorrow.
Previously, there was a race condition where the state of a job could change between when you last looked and when you went to update the database with the transition. Now the DAO object has a where clause to limit to only jobs that were in the (supposed) old state and will check the number of rows updated, raising a RuntimeError if they don't match up.

git-svn-id: svn+ssh://svn.cern.ch/reps/CMSDMWM/WMCore/trunk@4572 4525493e-7705-40b1-a816-d608a930855b