This is the preliminary check-in for the JobAccountant.
It still lacks the particulars of the closeout and tarup stages, since these appear
not to be particularly well documented.

It also depends on an ad-hoc method (not checked in) for getting runs out of JSON.  Hopefully
this will be fixed somewhat independently, and will not require changes to JobAccountant.

git-svn-id: svn+ssh://svn.cern.ch/reps/CMSDMWM/WMCore/trunk@4685 4525493e-7705-40b1-a816-d608a930855b