First, you shouldn't use this because I think it's broken, but I wanted to make a check-in while it at least passed the tests
Thsi is the first implementation of JobSubmitter using the WMRuntime package (using WMRuntime emulators to run jobs instead of
silly test jobs), and bringing back the new format JobReports.  I wouldn't trust this if I were you, but it does appear to run
everything properly on the worker node.  This is not yet integrated with other components, and will not run in integration
testing at this time.

git-svn-id: svn+ssh://svn.cern.ch/reps/CMSDMWM/WMCore/trunk@6225 4525493e-7705-40b1-a816-d608a930855b