This version of the popular AccountantWorker class is the third in the "Hey!  Let's move everything to bulk" series.  Added features in this version include the removal of any dependence on file ID for WMBS files, allowing for the removal of any exists() calls.  It also includes the centralization of the WMBS calls at the end of the processing, and even more call elimination.
The next edition, AccountantWorker Vista, is expected to include
updated performance that will move the code to a point where it will be
almost fast enough to use.  Improvements are expected to include either
a streamlined DBS file parentage system, or code that automatically
dispatches assassins to elminate redneck files.

git-svn-id: svn+ssh://svn.cern.ch/reps/CMSDMWM/WMCore/trunk@6695 4525493e-7705-40b1-a816-d608a930855b