Also known as AccountantWorker Vista Home Basic:
This version upgrades the AccountantWorker to run over multiple
jobs in one go instead of just one at a time.  It does not make
any additional bulk calls and should perform no differently then
the previous check-in, but has the possibility of being improved
on in the future.  It is being checked in now to provide a stable
multi-job version for future tests.

git-svn-id: svn+ssh://svn.cern.ch/reps/CMSDMWM/WMCore/trunk@6703 4525493e-7705-40b1-a816-d608a930855b