We call this one AccountantWorker Vista Home Premium
This is the first multi-job edition of the AccountantWorker.  It
takes advantage of the abiliy to commit in bulk at the end of the
process, allowing you to reduce the number of commit statements
to one per table modified per N job reports sent to the process.
This provides some significant speed increases, and makes the
handling of DBS file parentage the largest time hurdle remaining.

git-svn-id: svn+ssh://svn.cern.ch/reps/CMSDMWM/WMCore/trunk@6706 4525493e-7705-40b1-a816-d608a930855b