There are two changes in this version
One is that the code that actually runs the unpacker is moved to a function.
This makes testing it much easier.

The second is that the tarfile chunk has been put back in.  This makes it
impossible to run unpacker properly in python2.4, but tarfile resolves
some problems with untarring in general.  Since we have the rpm now, I
figured we should go to 2.6

git-svn-id: svn+ssh://svn.cern.ch/reps/CMSDMWM/WMCore/trunk@6752 4525493e-7705-40b1-a816-d608a930855b