A small adjustment to the Service.py after last Friday's fun and games which should alleviate issues caused by a service being down for a short period.
Previously the Service class did dumb caching - throwing an exception if a service is unreachable This change to getData checks an internal flag (self['usestalecache'], which defaults to False, e.g. current behaviour) and if the request raises an exception it'll return the previous version of the cache file if the flag is set. To avoid masking a long term problem with the remote service the cache file is marked as 'dead' when it is five times older than the cache duration - this would give operators the possibility to extend a cache file's lifetime using touch in a dire emergency.

self['usestalecache'] should be set to True in the constructor dict where services do not change frequently (for instance SiteDB naming convention API's would be a good contender for using it, a status call to the T0 DAS would likely not be).

Also, verb now defaults to GET which fixes up failing tests, the verb wasn't set in the tests.

git-svn-id: svn+ssh://svn.cern.ch/reps/CMSDMWM/WMCore/trunk@7574 4525493e-7705-40b1-a816-d608a930855b