Okay, this is just getting ridiculous.
This is my 4th rewrite of the DBSUpload system.  This is entirey different;
it handles the blocks locally and thus has no need to depend on DBS to do
either file grouping or file bookkeeping.  It also uses subprocesses to
handle individual Dataset-Algo pairs in order to run queries to DBS and
to our local database in parallel.

This has gotten to so many new versions that I'm adopting codenames.  This
one is called Marmot, because apparently Penguin and Bunny are already taken.
This requires the DBSInterface and the DBSUploadWorker to run.

I gotta lie down...

git-svn-id: svn+ssh://svn.cern.ch/reps/CMSDMWM/WMCore/trunk@7735 4525493e-7705-40b1-a816-d608a930855b