Track progress from WMBS in WorkQueue. Report/Record up chain and update ReqMgr.
When reporting a workqueue element status to the parent obtain subscription status from WMBS, collate it and send to the parent. Progress is stored and so available to WorkQueue monitoring.

Report status back to ReqMgr providing request status and progress metrics.

Not tested on Oracle or MySQL or with real ReqMgr.

git-svn-id: svn+ssh://svn.cern.ch/reps/CMSDMWM/WMCore/trunk@7865 4525493e-7705-40b1-a816-d608a930855b