Today's third revision of the JobSubmitterPoller
This one uses a new DAO to handle finding and loading jobs in one step.

It also periodically polls ResourceControl to update the status of running jobs.  This
happens by default once every 10,000 jobs, although this is a tunable parameter.  This
allows us to function under very long loads, as it allows us to benefit from the extra
slots generated when other jobs finish.

Unfortunately, it's also difficult to test, so it needs some thorough working over.

git-svn-id: svn+ssh://svn.cern.ch/reps/CMSDMWM/WMCore/trunk@8107 4525493e-7705-40b1-a816-d608a930855b