Fixes to allow global to split by block.
o Major change is to allow splitters to be given the data they should split on directly rather than talking it from the wmspec. This is needed for a local queue if the global split by block to avoid duplicate work being injected into wmbs. This means elements in lower queues which have been split by block before can still all refer to the same wmspec (as far as wmbs is concerned).

git-svn-id: svn+ssh://svn.cern.ch/reps/CMSDMWM/WMCore/trunk@8279 4525493e-7705-40b1-a816-d608a930855b