This contains all the hooks for coroutine-based JobSplitting.
Note that these hooks are NOT mandatory; all previous functionality is maintained.  However,
if you activate the right hooks you will begin pulling jobs out in small groups.

This is done by calling the open() method, which opens the result proxies to the database,
and flags the code to pull jobs using the proxies instead of simple DAOs when you call
sortByLocation().  Once this is done it will pull in a number of new files each iteration
equivalent to the limit you passed in at the beginning (by setting limit in __init__).  If
you don't want to set a global limit, or you want to override it, you need to call
the JobFactory with an argument 'file_load_limit'.

This code is built to work with the new JobCreator (Taizong) upgrade, and should be
installed with the other packages.

git-svn-id: svn+ssh://svn.cern.ch/reps/CMSDMWM/WMCore/trunk@8362 4525493e-7705-40b1-a816-d608a930855b