This is the long-awaited (by some) upgrade to the JobCreator, package codename: Taizong
It deals with the problem of the JobCreator having to pause while working through a large
chunk of jobs by using generators to iterate through SQL, pulling out a chunk of jobs with
each iteration, and then running those all the way through.  This should mean that instead
of everything holding while it processes all jobs, it should slowly move jobs forward as
it proceeds.  This requires the JobCreator to handle subscriptions in an iterative manner.

In order to specify the number of files to be pulled out at a time in the JobSplitting,
you want to work with the fileLoadLimit config value for the JobSubmitter.  By default it
pulls out a hundred records at a time.  For a particular workflow this can be overridden
by messing with the workflow splitting params, although I'm not sure you want to do this.

It should work with everything, but is not well tested.  Use at your own risk.

git-svn-id: svn+ssh://svn.cern.ch/reps/CMSDMWM/WMCore/trunk@8364 4525493e-7705-40b1-a816-d608a930855b