Move wmspec insertion into transaction. fix #115.
Previously duplicate work check was before splitting and in separate
transaction thus a failure in splitting would leave the wmspec in the
database. Subsequent attempts to obtain the work would result in the
spurious duplicate work message.

Move check into transaction and rollback on duplication, this forces the
check after the split.

https://svnweb.cern.ch/trac/CMSDMWM/ticket/115

git-svn-id: svn+ssh://svn.cern.ch/reps/CMSDMWM/WMCore/trunk@8533 4525493e-7705-40b1-a816-d608a930855b