Ensure request cancelation propagates to WMBS.
o Problem was that WorkQueue REST interface was told not to interact
with WMBS (i.e. for getWork() call) & therefore didn't pass on the cancel.

o Soloution was to use 2 flags, one to tell the WorkQueue not to inject
work into wmbs; and the other to control whether it can modify exisiting
WMBS subscriptions.

Fixes #813

From: Seangchan Ryu <sryu@fnal.gov>
Signed-off-by: Stuart Wakefield <stuart.wakefield@imperial.ac.uk>


git-svn-id: svn+ssh://svn.cern.ch/reps/CMSDMWM/WMCore/trunk@11148 4525493e-7705-40b1-a816-d608a930855b