Fix workqueue elememt priority calculation.
Couchdb only evaluates the view function for an element once and then
caches the result. The previous calculation used the difference between
insertion time and current time to increase the prioirty of old
elements. Thus the priority would not change as expected and would be
wrong for subsequent calls.

Fix by changing priority calculation to be a static number which does
not need to be re-evaluated each time.

Fixes #2112

From: Stuart Wakefield <stuart.wakefield@imperial.ac.uk>
Signed-off-by: Stuart Wakefield <stuart.wakefield@imperial.ac.uk>


git-svn-id: svn+ssh://svn.cern.ch/reps/CMSDMWM/WMCore/trunk@13873 4525493e-7705-40b1-a816-d608a930855b