Subject: Build sphinx documentation.
This adds very rudimentary documentation. If the requested system does
not have a separate directory, uses the general 'wmcore' one instead.
In future it would be expected to have custom documentation for the
systems.

The main benefit of the documentation at the moment is that includes
the full module index with source code for the selected systems.

From: Lassi Tuura <lat@cern.ch>


git-svn-id: svn+ssh://svn.cern.ch/reps/CMSDMWM/WMCore/trunk@14767 4525493e-7705-40b1-a816-d608a930855b