Split module documentation into individual files.
Default generate_modules.py produces documentation for all modules of
a package into a single file. That is overwhelming amount in a single
HTML page, split the documentation for each module into a separate
file.

This effectively imports the corresponding change from SiteDB, but
with tweaks to make it work with WMCore's more nested structure.