Updating new storage/stageout plugins
This commit fixes StageIn so it does the right thing, and additionally
moves untested plugins to the Examples directory. The new stageout code
will fallback to the old stageout code if the plugin doesn't exist.

I tested this at Vandy with the vandy plugin, so the framework itself
works for stagein/out/delete