Handle files with multiple runs in Harvest
Modifiy the Harvest JobSplitter algorithm to handle multiple
runs in a file, it adds the run/lumi information to the job mask.

Note that this patch doesn't remove the special do-not-mix-runs
behavior of the agent, that will go in another patch which
will need https://github.com/dmwm/WMCore/pull/39 to
be applied.