change taskarchiver subscription delete order
Major functional change is changing the order in which subscription
ids are returned from the GetFinishesSubscription DAO. If multiple
subscriptions are found in the same query, the ones with the higher
ids are deleted first. This fixes a bug in the Tier0 where a working
directory for subscription 2 was a subdir of the working directory
for subscription 1 and they were deleted out-of-order.

In the process of tracking this down I also rewrote and simplified
the query in the DAO.