Protect JobTracker from bad pickles
Happened on vocms201, a job was failed due to timeout but a zero size job report
pickle was stored in the cached directory. The JobTracker then constantly
went down when trying to load said pickle.

The fix catches any exception when loading a FWJR after a timeout and
generates a new valid FWJR stating that the job failed due to a timeout
and logs the error in the JobTracker log.