fix some issues in the fwj report
When calling the constructor of the Report class with
a step name, it adds the step and initializes self.report
to point at the data for that step. Some other methods
to modify the report then work off self.report.

When unpersisting the report (ie. loading data from
a file and overwriting the internal data), self.report
will now no longer point at the internal data structures.
The modification moethds will still work, they just
modify some data that is no longer attached to the
class and will not be written out when the report is
persistet again.

Fix this by allowing to pass a step name to the unpersist
methos, it then initializes self.report to point at the
data for that step after loading the file. Also fix a
few places where the Report class is used to not pass
a stepname to the constructor if it's not used.