have TaskArchiver accept missing working directory
Currently, the TaskArchiver will print an error message and send
an alert if it cannot delete the working directory for a task due
to that working directory not existing.

If the input for a task never gets any events, the task itself is
never initialized by the JobCreator and therefor it's working
directory is never created in the first place. This is normal
for these cases and not an error.

Downgrade the error message to a debug message and remove the alert.

While modifying this code anyways, make the consistency check more
specific: If the working directory does not exist, handle it in
the described way. If it exists, but is not a directory, bail out
with an exception because at that point something is screwed up.