fixes in AlcaHarvest runtime code to make it play nice with the JobAccountant
Files in the FJR are organized under attributes named after
the output module label. Each file itself needs to have the
output module label specified again in it's attributes
though, otherwise the JobAccountant will register the
file in WMBS, but not register it in any of the filesets
it needs to go to. So when faking a Sqlite output module
with files in it, I need to add the module label as
file attribute as well.