Add the "New" state to the LSF plugin map
Even though the initial state of BossAir can be configured
as PEND for agents with the LSF plugin configured, if
we want an agent that uses both condor and LSF plugins that option
would not work anymore. Therefore the LSF plugin
should map the New state to Pending as well.