Start finishing subscriptions from the top
Before the latest changes the subscriptions were finished from
the bottom because they were deleted on the spot. However
now that the deletions only take place at the very end when everything
is finished we can actually remove the requirement that the child
subscriptions are finished to mark the parent as finished. This will
allow us to have a view of the progress of a workflow in terms of its
finished subscriptions, e.g. a workflow finished the DataProcessing subscription
and now it's just waiting on the merge. Also the more we can skim from this
dreaded query the better.