Store phedex subscription information in workloads
Modify the WMTask and WMWorkload classes to store subscription
information for each outputModule/dataset in each task.

This can be specified during assign, the reqmgr web interface now shows
another set of site lists for subscription configuration. The subscription
configuration can be totally unrelated to the site whitelists.

The options that can be configured for a subscription are:

- Custodial sites
- Non custodial sites
- Sites where the request can be auto approved
- Priority

The option to customize subscription information per task is not exposed in the ReqMgr
for now is only intended for the Tier-0 use case.

This a ReqMgr-only patch