Improve the PhEDEx subscriber
The PhEDEx subscriber component now uses the spec information
to make the subscriptions regardless of where the files ended up.

The specs are copied to a directory in the JobAccountant and its
path is stored in the DBSBuffer workflow table under the column "spec".

The PhEDEx subscriber can operate in 2 modes:

- Unsafe: Subscribes custodial as Move as soon as files appear in DBSBuffer
- Safe: Subscribes custodial as Replica first, and when the workflow is complete as Move

Also implemented a PhEDEx subscriber unit test using the PhEDEx emulator and fixed
the broken JobAccountant_t test.

This is a client-only patch, if an agent is updated but the reqmgr doesn't have
the previous commit then the PhEDEx subscriber won't do anything. But it is stable,
i.e. it won't crash.