Make TaskChain more configurable
Changes with this patch:

- Allow different parameters which were constant throughout
the workflow to be configurable per task, this are usually set
on the StdBase instance. Add a decorator to modify
this instance before any task is setup and restore it
afterwards.

- Cleanup the unit test for the TaskChain spec and test
the new functionality.

- Allow that some tasks' output be only in the unmerged
area only for the use of later tasks but not propagated to
DBS.

- Change the dashboard activity to relval, makes more sense
than having some under production and others under processing.

This is a pure ReqMgr patch, no changes needed on the agent.

Fixes #4138, #4117, #4116, #4112