Modify LFN counters in extension workflows
Add the initial_lfn_counter parameter to the splitting of the
production and merge tasks in MonteCarlo and LHEStepZero workflows
if the FirstEvent/Lumi is different from 1. The initial_lfn_counter
is calculated as what the last previous job count should have been
based on the current events_per_job and FirstEvent.

Fixes #4151