Fixing giant log files/memory leaks
WMException was printing traceback.format_exc(), which spits out all the
information for an exception. In the case of nested exceptions, this
would cause the size of the error message would grow geometrically.
Also, WMException was looking for member variables it didn't have, so it
would end up failing in __init__ (classy)