implement alternative fileset closing policy
Normally the JobArchiver will only start closing filesets if
the corresponding workflow has been marked as injected. Implement
an alternative policy where fileset closing ignores the injection
status of the workflow and only depends on whether the top level
fileset is closed or not. This is for workflows where the
chain of subscriptions only goes down, that is, we do not have
multiple different top level filesets and subscriptions feeding
into the same lower level fileset/subscription.