Adding support to write files into sub-user space
Current (CRAB2) LFNs support LFNs that look like

/store/user/lpctlbsm/meloam/blah/blah

Where lpctlbsm is the group name. Not only is it really conveniant
to be able to keep an idea of whose files are which in the group storage
(i.e. du -ckhs or dcdu -ckhs), but end-users are going to expect their
files to end up on the same place, regardless of what tool was used to
produce them