Bugfix for PrimaryDataset handling in RelVal
The PrimaryDataset argument wasn't really suited to be included in the
ParameterStorage decorator used to modify global parameters according
to task-level specifications. Since the PrimaryDataset
is not really a global parameter and always a task one, changed this support
for PrimaryDataset in order to make this actually work.

Added unit test verification of the output datasets to catch these errors.

Fixes #4225