Tier0 Plugin implementation
The Tier0 plugin for the AnalyticsDataCollector components
calculates the state of Tier-0 workflows based on subscription
information found in WMBS and updates the request document
in central WMStats with this status.

Additionally this patch fixes some bugs with PluginInterface implementation
patch and renames it from PlugIn to Plugin to conform with the same
name as the BossAir package (PlugIn is a bit harder to remember).

Fix the AnalyticsDataCollector unittest and removed it from the
allowed failing tests.

This requires #4227 to work, it also fixes the unit test breaks from it.

Fixes #4264