Implement override for PhEDEx injections
Due to the special setup of the T0/T2 at CERN
and the Bulk/Express processing in the Tier-0. We need
the ability to specify where the files should be registered
to in PhEDEx. This is done per spec, and all files
associated with the workflows from that spec
will be registered to the given site, regardless of the SE
available in the dbsbuffer table or custodial site
entry in dbsbuffer.

Also fix a unit teste in WMWorkload_t