Implement Pending jobs per task/site functionality
Add a new column to the rc_threshold table, this indicates the number of pending
jobs to keep for an individual task.

Modify all the concerned DAOs and the JobSubmitter.

Rewrote the JobSubmitterPoller unit tests to make them all non-integration,
break the dependecy of having condor available and add tests for the different
mechanics happening in it such as threshold management, site status,
site whitelists/blacklists. Also includes a performance test.

Finally add a special option to setup the Tier-0 thresholds in one line from the
resource control script.