Include REST API for missing lumis in the summary
Add a show function which works as a REST API
to get the missing lumis for a completed workflow from
the workload summary.

URL is ugly at the moment:

/_design/WorkloadSummary/_show/lumiMissingByWorkflow/<workflowName>?outputDataset=<string>

The outputDataset parameter is optional and is used to filter the output. The format of the output
is:

{<outputDataset:string> : { <run:string> : [ [<startLumi:int>, <endLumi:init>], ...] , ... }, ...}

Note that for summaries produced by agents without the TaskArchiver changes in this patch, the outputDataset won't
be returned as a key, instead it will be the task where the lumis were lost.

Fixes #4216