Implement summary of failure per site
Include two type of histograms with the number of failed jobs
per site. One given per site at a workflow level, and other
per step and exit code.

This is included in the summary using the new Histogram data structure
that allows generic histograms to be built with a simple interface.
This strutcture takes care of statistic calculation and generation
of a dictionary suitable for couchdb, also is meant to provide
support for multiple agents which will require merging summaries.

Fixes #4213