Fix Reduce function for outputByWorkflowName
It was generating a growing object in the output,
although it would only grow to size 2 at any given time.
Couchdb reduce_limit check throws an overflow
error on this function, changing the type to an array
fixes the problem. Added an object variable
that only lives in the reduce function to
keep track of duplicates.

Changed the TaskArchiver accordingly.