Avoid crashes on custodial subscription to non-MSS
The PhEDEx subscriber must check if a custodial
subscription will be made to a MSS node before doing it
otherwise PhEDEx will reject it and the poller will crash.

If a custodial subscription is made to a site without MSS
phedex node (no tape backend) then it defaults to a non-custodial
subscription.