Fix EOS StageIn at CERN
Plugin was using rfcp to stageIn files from EOS at CERN.
Check properly if remotePFN is EOS, instead of targetPFN
as in stageIn the targetPFN will be the local.

Also fix deletions, in stageIn case it was going to
delete the remotePFN instead of the faulty local
copy. Include deletions for failed CASTOR stage-outs.

Finally, only create directories on remote PFNs.